changed pause to accept a double
git-svn-id: http://svn.warlock.cc/svn/warlock/warlock2/trunk@1215 bb130017-240f-0410-9613-9c2c827820fb

+++ core/cc.warlock.core.script/src/main/cc/warlock/core/script/IScriptCommands.java
@@ -18,7 +18,7 @@ public interface IScriptCommands {
 	
 	public Match matchWait ();
 	
-	public void pause (int seconds);
+	public void pause (double seconds);
 	
 	public void waitForPrompt ();
